Makefile: Build in a -ffreestanding environment
Tell GCC that we're compiling in a freestanding environment, e.g. one
in which standard libraries do not exist and in which GCC should not
use builtin functions. This allows us to reuse some of the common
function names, such as exit(), but with our own implementation.

Turning off builtin functions also stops warnings like the following
from being emitted,

efilinux.h:187:1: warning: conflicting types for built-in function ‘exit’

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>