efilinux: Pass correct size to exit()
The @size argument to exit() should be the size, in bytes, of
@reason. Previously we were passing whatever value was left in 'size'
in efi_main(). We should instead be passing ERROR_STRING_LENGTH
because that is the size of the allocation holding @reason.

Also, cleanup a typo in the documentation for exit() where the
explanation of the @size argument referred to a nonexistent @data
argument.

Signed-off-by: Matt Fleming <matt.fleming@linux.intel.com>