bzimage: Fix typecast bug on i386
Before 32-bit support we were playing fast and loose with casts. The
problem was that we were casting away legimate warnings, such as the
compiler telling us that the value it was storing into a variable to
was too big to fit.

emalloc() was a prime example, in various places we passed a pointer
to a pointer as the third argument to emalloc which has type
EFI_PHYSICALL_ADDRESS *, where EFI_PHYSICAL_ADDRESS is a 64-bit type,
like so,

	emalloc(UINTN size, UINTN align, EFI_PHYSICAL_ADDRESS *addr)

Then inside emalloc() we do,

	*addr = aligned;

On 64-bit this is fine because,

	sizeof(void *) == sizeof(EFI_PHYSICAL_ADDRESS)

but if we have a 32-bit pointer we attempt to store a 64-bit value
into a 32-bit data type. This led to stack corruption when we passed
the address of a pointer on the stack to emalloc().

Signed-off-by: Matt Fleming <matt.fleming@intel.com>