fs: Don't copy junk from locate_handle() buffer
Currently if we have more than one handle that supports
FileSystemProtocol we'll copy junk from 'fs_devices' when passing the
arguments to handle_protocol(). This is because we're using an
incorrect data type for the buffer filled out by locate_handle() - it
returns an array of EFI_HANDLE, not an array of struct fs_device's.

Instead use a temporary buffer to store the list of EFI_HANDLE's and
copy them individually to fs_devices[].handle.

Also, this is a good opportunity to clean up the error path in
fs_init() so no longer leak open handles and 'fs_devices'.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>