efilinux: malloc() size is too small for filename
We should be using sizeof() to figure out how large the buffer needs
to be because the buffer does not hold an array of bytes, rather it's
an array of CHAR16, which is actually 2 bytes. While I'm here cleanup
the error path so that we free() '*name' and '*cmdline' in case of
failure.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>