Makefile: Make cross-building possible
Add support for building ia32 efilinux on an x86-64
host. Cross-building can be turned on by specifying $ARCH when
building efilinux, e.g. (assuming an x86-64 host)

	$ make ARCH=ia32

Also, delete $MACHINE from the Makefile and use $ARCH instead as ARCH
seems to be more commonly used for specifying the target architecture;
gnu-efi uses $ARCH for example.

Note that $(CFLAGS) is now passed to $(CC) when looking up the path
for libgcc.a,

	$(shell $(CC) $(CFLAGS) -print-libgcc-file-name)

This change is required because if we're cross-building $(CFLAGS) will
include -m32, which makes -print-libgcc-file-name return the filename
of the 32-bit version of libgcc.a, not the host's version.

Reported-by: Darren Hart <dvhart@linux.intel.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>