Find crt0 and the LD script automatically
gnuefi doesn't install crt0 and the ld script to a gnuefi directory by
default, but some distributions stick it there. Use find via a shell
command to take the last matching file for each of CRT0 and LDSCRIPT.

This avoids forcing the caller to duplicate the script generating the
ARCH variable used throughout the Makefile.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
CC: Matt Fleming <matt.fleming@intel.com>