fs: Don't print message if we fail to open a file
Now that we have config file support failing to open a fail isn't
worthy of a warning. Many users won't be using a config file at all
and printing,

	"Unable to open file <filename>"

everytime efilinux is executed is useless and annoying. Leave it to
the caller of file_open() to print any messages about failing to open
files.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>