Makefile: Do not include symbol table in efilinux.efi
objcopy will leave the symbol table intact when creating an EFI
executable from the ELF shared object, which is unnecessary, but more
importantly the existence of the symbol table causes some EFI hashing
tools to fail when signing the executable for secure boot because the
symbol table is considered to be garbage.

Strip all symbol table and relocation information when creating the
EFI executable.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>